requirements.txt
streamlit
transformers
torch
numpy
pandas
Pillow